+++ src/main/java/insane96mcp/enhancedai/modules/Modules.java
@@ -13,6 +13,7 @@ public class Modules {
 	public static Module creeper;
 	public static Module drowned;
 	public static Module ghast;
+	public static Module shulker;
 	public static Module golem;
 	public static Module enderman;
 	public static Module skeleton;
@@ -28,6 +29,7 @@ public class Modules {
 		creeper = Module.Builder.create(Ids.CREEPER, "Creepers", ModConfig.Type.COMMON, Config.builder).build();
 		drowned = Module.Builder.create(Ids.DROWNED, "Drowneds", ModConfig.Type.COMMON, Config.builder).build();
 		ghast = Module.Builder.create(Ids.GHAST, "Ghasts", ModConfig.Type.COMMON, Config.builder).build();
+		shulker = Module.Builder.create(Ids.SHULKER, "Shulkers", ModConfig.Type.COMMON, Config.builder).build();
 		golem = Module.Builder.create(Ids.PETS, "Pets", ModConfig.Type.COMMON, Config.builder).build();
 		enderman = Module.Builder.create(Ids.ENDERMAN, "Endermen", ModConfig.Type.COMMON, Config.builder).build();
 		skeleton = Module.Builder.create(Ids.SKELETON, "Skeletons", ModConfig.Type.COMMON, Config.builder).build();
@@ -44,6 +46,7 @@ public class Modules {
 		public static final String CREEPER = EnhancedAI.RESOURCE_PREFIX + "creeper";
 		public static final String DROWNED = EnhancedAI.RESOURCE_PREFIX + "drowned";
 		public static final String GHAST = EnhancedAI.RESOURCE_PREFIX + "ghast";
+		public static final String SHULKER = EnhancedAI.RESOURCE_PREFIX + "shulker";
 		public static final String PETS = EnhancedAI.RESOURCE_PREFIX + "pets";
 		public static final String ENDERMAN = EnhancedAI.RESOURCE_PREFIX + "enderman";
 		public static final String SKELETON = EnhancedAI.RESOURCE_PREFIX + "skeleton";